+++ CS_461_Programming_and_Writing_Project/CS_461_Programming_and_Writing_Project/Point.h
@@ -20,7 +20,7 @@ public:
     Point(); 
 
     // constructor 
-    Point(int a, int b); 
+    Point(double a, double b); 
 
     // copy constructor
     Point(const Point& p);
